requirements.txt
docling==2.30.0
fastapi==0.128.0
gradio==6.2.0
httpx[socks]==0.28.1
openai==2.14.0
Requests==2.32.5
python-dotenv>=1.0.0
PyYAML>=6.0
google-generativeai>=0.8.0
hf_xet>=1.0.0